+++ dataprovider/espn/mma/example_test.go
@@ -4,16 +4,18 @@ import (
 	"encoding/json"
 	"fmt"
 	"log"
+	"time"
 
 	"github.com/lightning-dabbler/sportscrape/dataprovider/espn/mma"
 	"github.com/lightning-dabbler/sportscrape/runner"
+	"github.com/lightning-dabbler/sportscrape/scraper"
 )
 
 // Example for mma.ESPNMMAMatchupScraper
 func ExampleESPNMMAMatchupScraper() {
 	matchupRunner := runner.NewMatchupRunner(
 		runner.MatchupRunnerScraper(
-			mma.ESPNMMAMatchupScraper{Year: "2024", League: "ufc"},
+			mma.ESPNMMAMatchupScraper{Year: "2024", League: "ufc", BaseScraper: scraper.BaseScraper{Timeout: time.Second * 10}},
 		),
 	)
 
@@ -35,7 +37,7 @@ func ExampleESPNMMAFightDetailsScraper() {
 
 	matchupRunner := runner.NewMatchupRunner(
 		runner.MatchupRunnerScraper(
-			mma.ESPNMMAMatchupScraper{Year: "2024", League: "ufc"},
+			mma.ESPNMMAMatchupScraper{Year: "2024", League: "ufc", BaseScraper: scraper.BaseScraper{Timeout: time.Second * 10}},
 		),
 	)
 
@@ -45,7 +47,7 @@ func ExampleESPNMMAFightDetailsScraper() {
 	}
 
 	eventRunner := runner.NewEventDataRunner(
-		runner.EventDataRunnerScraper(mma.ESPNMMAFightDetailsScraper{}),
+		runner.EventDataRunnerScraper(mma.ESPNMMAFightDetailsScraper{League: "ufc", BaseScraper: scraper.BaseScraper{Timeout: 10 * time.Second}}),
 	)
 
 	fightDetails, err := eventRunner.Run(result[0:10]...)